+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 ### Added
 - Solving stats methods for number of nodes, time, and number of lp iterations. 
+- Branching rule plugin. 
 ### Fixed
 ### Changed
 ### Removed